Add GitHub Actions (#1)
Add Actions for building the thesis, publishing a PDF preview on PullRequests
or to GitHub Pages if pushed to main. Also add Actions for cleaning preview
branches after the PR is merged and a preview message on the PRs. Create
a requirements-ubuntu.txt for all the Ubuntu packages needed to compile the
thesis.