Configure headers and footers (#3)
Add only page number on the bottom of every page of the frontmatter. Add
a header on the pages of the mainmatter with page number on the outer side,
chapter title on even pages and section title on odd pages. Remove any header
and footer from chapter pages and from blank pages that might appear before
them.